+++ src/gpu.rs
@@ -1,9 +1,6 @@
 use std::{marker::PhantomData, sync::Mutex, cell::RefCell};
-
 use wgpu::util::{DeviceExt, BufferInitDescriptor};
-
 use crate::{utils::Byteable, math::Vec3};
-
 
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
@@ -30,11 +27,11 @@ impl<T> GPUVec3<T> where T : Byteable
     }
 }
 
-impl<T> Into<GPUVec3<T>> for Vec3<T> where T : Byteable
+impl<T> From<Vec3<T>> for GPUVec3<T> where T : Byteable
 {
-    fn into(self) -> GPUVec3<T> 
+    fn from(value: Vec3<T>) -> Self 
     {
-        GPUVec3::new(self.x, self.y, self.z)
+        GPUVec3::new(value.x, value.y, value.z)
     }
 }
 
@@ -69,8 +66,28 @@ impl<T> GBuffer<T> where T : Byteable
         }
     }
 
+    pub fn new_empty(capacity: u64, usage: wgpu::BufferUsages, device: &wgpu::Device, label: Option<&str>) -> Self
+    {
+        let handle = device.create_buffer(&wgpu::BufferDescriptor {
+            label,
+            size: capacity * std::mem::size_of::<T>() as u64,
+            usage,
+            mapped_at_creation: false
+        });
+
+        Self
+        {
+            length: capacity,
+            capacity,
+            handle,
+            usage,
+            phantom: PhantomData {}
+        }
+    }
+
     pub fn length(&self) -> u64 { self.length }
     pub fn capacity(&self) -> u64 { self.capacity }
+    pub fn size(&self) -> u64 { self.length * std::mem::size_of::<T>() as u64 }
 
     pub fn enqueue_set(&mut self, data: &[T], queue: &wgpu::Queue)
     {
@@ -80,7 +97,7 @@ impl<T> GBuffer<T> where T : Byteable
 
     pub fn slice(&self, start: u64, end: u64) -> wgpu::BufferSlice
     {
-        assert!(start <= end, "Start index must be less than or equel to the end index");
+        assert!(start <= end, "Start index must be less than or equal to the end index");
         assert!(end < self.length(), "Slice is larger than the contained data");
         self.handle.slice(start..end)
     }
@@ -88,6 +105,11 @@ impl<T> GBuffer<T> where T : Byteable
     pub fn slice_all(&self) -> wgpu::BufferSlice
     {
         self.handle.slice(0..self.length())
+    }
+
+    pub fn read(&self, device: &wgpu::Device) -> Vec<T>
+    {
+        pollster::block_on(self.read_async(device))
     }
 
     pub async fn read_async(&self, device: &wgpu::Device) -> Vec<T>
@@ -120,11 +142,19 @@ impl<T> GBuffer<T> where T : Byteable
 
     pub fn copy(&self, dest: &mut GBuffer<T>, command_encoder: &mut wgpu::CommandEncoder)
     {
-        
+        assert!(dest.capacity >= self.length, "Destination buffer capacity not large enough");
+        command_encoder.copy_buffer_to_buffer(&self.handle, 0, &dest.handle, 0, self.size());
+        dest.length = self.length;
     }
 
     pub fn as_entire_binding(&self) -> wgpu::BindingResource
     {
         self.handle.as_entire_binding()
     }
+}
+
+pub struct ShaderInfo<'a>
+{
+    pub entry_point: &'a str,
+    pub source: &'a str
 }
