+++ src/voxel/world_gen.rs
@@ -3,8 +3,9 @@ use cgmath::Zero;
 use wgpu::{PipelineLayoutDescriptor, BindGroupLayoutDescriptor, BindGroupLayoutEntry};
 use wgpu::util::DeviceExt;
 use crate::math::Vec3;
-use crate::gpu::GPUVec3;
+use crate::gpu::{GPUVec3, ShaderInfo};
 use crate::utils::Array3D;
+use crate::gpu::GBuffer;
 
 use super::Voxel;
 
@@ -14,10 +15,10 @@ pub struct VoxelGenerator
     queue: Arc<wgpu::Queue>,
 
     chunk_size: Vec3<u32>,
-    staging_buffer: wgpu::Buffer,
-    storage_buffer: wgpu::Buffer,
-    chunk_size_buffer: wgpu::Buffer,
-    chunk_pos_buffer: wgpu::Buffer,
+    staging_buffer: GBuffer<u32>,
+    storage_buffer: GBuffer<u32>,
+    chunk_size_buffer: GBuffer<GPUVec3<u32>>,
+    chunk_pos_buffer: GBuffer<GPUVec3<i32>>,
 
     bind_group: wgpu::BindGroup,
     compute_pipeline: wgpu::ComputePipeline,
@@ -25,47 +26,31 @@ pub struct VoxelGenerator
 
 impl VoxelGenerator
 {
-    pub fn new(chunk_size: Vec3<u32>, device: Arc<wgpu::Device>, queue: Arc<wgpu::Queue>) -> Self 
+    pub fn new(chunk_size: Vec3<u32>, device: Arc<wgpu::Device>, queue: Arc<wgpu::Queue>, shader_info: ShaderInfo) -> Self 
     {
         // Loads the shader from WGSL
         let cs_module = device.create_shader_module(wgpu::ShaderModuleDescriptor {
             label: None,
-            source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(include_str!("../shaders/test_compute.wgsl"))),
+            source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(shader_info.source)),
         });
 
-        let size = (std::mem::size_of::<u32>() * (chunk_size.x * chunk_size.y * chunk_size.z) as usize) as wgpu::BufferAddress;
+        let length = (chunk_size.x * chunk_size.y * chunk_size.z) as u64;
 
-        let staging_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-            label: None,
-            size,
-            usage: wgpu::BufferUsages::MAP_READ | wgpu::BufferUsages::COPY_DST,
-            mapped_at_creation: false,
-        });
+        let staging_buffer = GBuffer::<u32>::new_empty(length, wgpu::BufferUsages::MAP_READ | wgpu::BufferUsages::COPY_DST, &device, Some("Staging buffer"));
 
-        let storage_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Storage Buffer"),
-            contents: bytemuck::cast_slice(&vec![0 as u8; size as usize]),
-            usage: wgpu::BufferUsages::STORAGE
-                | wgpu::BufferUsages::COPY_DST
-                | wgpu::BufferUsages::COPY_SRC,
-        });
+        let storage_buffer_usage = wgpu::BufferUsages::STORAGE
+                                 | wgpu::BufferUsages::COPY_DST
+                                 | wgpu::BufferUsages::COPY_SRC;
 
-        let gpu_vec_size: GPUVec3<_> = chunk_size.into();
-        let chunk_size_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: None,
-            contents: bytemuck::cast_slice(&[gpu_vec_size]),
-            usage: wgpu::BufferUsages::UNIFORM 
-                 | wgpu::BufferUsages::COPY_DST
-                 | wgpu::BufferUsages::COPY_SRC
-        });
+        let storage_buffer = GBuffer::<u32>::new_empty(length, storage_buffer_usage, &device, Some("Storage buffer"));
 
-        let chunk_pos_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: None,
-            contents: bytemuck::cast_slice(&[GPUVec3::<i32>::new(0, 0, 0)]),
-            usage: wgpu::BufferUsages::UNIFORM 
-                 | wgpu::BufferUsages::COPY_DST
-                 | wgpu::BufferUsages::COPY_SRC
-        });
+        let uniform_usage = wgpu::BufferUsages::UNIFORM 
+                          | wgpu::BufferUsages::COPY_DST
+                          | wgpu::BufferUsages::COPY_SRC;
+
+        let chunk_size_buffer = GBuffer::new(&[chunk_size.into()], uniform_usage, &device, Some("Chunk size buffer"));
+
+        let chunk_pos_buffer = GBuffer::<GPUVec3<i32>>::new_empty(1, uniform_usage, &device, Some("Chunk position buffer"));
 
         let bind_group_layout = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
             label: None,
@@ -115,7 +100,7 @@ impl VoxelGenerator
             label: None,
             layout: Some(&compute_pipeline_layout),
             module: &cs_module,
-            entry_point: "main",
+            entry_point: shader_info.entry_point,
         });
 
         let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
@@ -153,17 +138,17 @@ impl VoxelGenerator
         }
     }
 
-    pub fn run(&self, chunk_pos: Vec3<i32>) -> Array3D<u32>
+    pub fn run(&mut self, chunk_pos: Vec3<i32>) -> Array3D<u32>
     {
         pollster::block_on(self.run_async(chunk_pos))
     }
 
-    pub async fn run_async(&self, chunk_pos: Vec3<i32>) -> Array3D<u32>
+    pub async fn run_async(&mut self, chunk_pos: Vec3<i32>) -> Array3D<u32>
     {
         let voxel_count = self.chunk_size.x * self.chunk_size.y * self.chunk_size.z;
         let size = (std::mem::size_of::<u32>() * voxel_count as usize) as wgpu::BufferAddress;
 
-        self.queue.write_buffer(&self.chunk_pos_buffer, 0, bytemuck::cast_slice(&[GPUVec3::new(chunk_pos.x, chunk_pos.y, chunk_pos.z)]));
+        self.chunk_pos_buffer.enqueue_set(&[chunk_pos.into()], &self.queue);
 
         let mut encoder = self.device.create_command_encoder(&wgpu::CommandEncoderDescriptor { label: None });
         {
@@ -176,31 +161,13 @@ impl VoxelGenerator
             compute_pass.insert_debug_marker("compute random numbers");
             compute_pass.dispatch_workgroups(self.chunk_size.x, self.chunk_size.y, self.chunk_size.z); // Number of cells to run, the (x,y,z) size of item being processed
         }
-        
-        encoder.copy_buffer_to_buffer(&self.storage_buffer, 0, &self.staging_buffer, 0, size);
+
+        self.storage_buffer.copy(&mut self.staging_buffer, &mut encoder);
 
         self.queue.submit(Some(encoder.finish()));
         
-        let buffer_slice = self.staging_buffer.slice(..);
-        let (sender, receiver) = futures_intrusive::channel::shared::oneshot_channel();
-        buffer_slice.map_async(wgpu::MapMode::Read, move |v| sender.send(v).unwrap());
-
-        self.device.poll(wgpu::Maintain::Wait);
-
-        let result: Vec<u32> = if let Some(Ok(())) = receiver.receive().await {
-            let data = buffer_slice.get_mapped_range();
-            let result = bytemuck::cast_slice(&data).to_vec();
-
-            drop(data);
-            self.staging_buffer.unmap();
-
-            result
-        } 
-        else 
-        {
-            panic!("failed to run compute on gpu!")
-        };
-
+        
+        let result = self.staging_buffer.read(&self.device);
         Array3D::from_vec(self.chunk_size.x as usize, self.chunk_size.y as usize, self.chunk_size.z as usize, result)
     }
 }
